+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  * Improved ellipse arc to cubic bezier conversion
  * Fixed splitting beziers with less than 2 points
  * Added more lottie objects to the model
+ * Customizable export width/height
 
 ## 0.6.10 - 2021-09-18
  * Glaxnimate compatibility